requirements.txt
google-genai >= 1.16.1
streamlit >= 1.45.1
pillow  >= 11.2.1